requirements.txt
fastapi-cache2 # Caching
prometheus-fastapi-instrumentator # Monitoring
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv